Split flask cli commands out of api.py into cli.py
api.py now only holds HTTP concerns (routes, generate_fw, optional_auth).
cli.py holds import_json, submit_firmware, the shared _upsert_firmware
helper, and VALID_FW_KINDS, with its own init_app that registers the
click commands. cohorts/__init__.py calls both init_api and init_cli.
